split MyMap into cpp and hpp
